% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exec.R
\name{exec_aws_cli}
\alias{exec_aws_cli}
\title{Execute AWS CLI command, sub-command, and args.}
\usage{
exec_aws_cli(cmd, subcmd, ...)
}
\value{
A list with entries [int] \code{status}, [string] \code{stderr}, and [string] \code{stdout}.
}
