% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list2args.R
\name{dots2args}
\alias{dots2args}
\alias{list2args}
\title{Convert a list to AWS CLI args}
\usage{
dots2args(...)

list2args(l)
}
\arguments{
\item{l}{A list of arguments to be mapped to AWS CLI args.
#' @details
  Unnamed list entries are inserted into the system call as (positional) "<val>" arguments, i.e. with no argument-specifying hyphens.
  Named list entries are inserted into the system call with the "--<name> <val>" pattern.
  Any named list entries with a \code{NULL} value are inserted into the system call as "--<name>" ut with no additional value, i.e. those arguments typically specify named flags.
  The AWS CLI call's arguments are built in the order in which they appear in the list, so if an argument appears more than once in the list, there'll be multiple such args in the system call, too (in the same order in which they appear in the list).}
}
\examples{
list2args(list("unnamed", "named" = "value", "flag" = NULL))
}
