% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3api.R
\name{s3api_put_object}
\alias{s3api_put_object}
\title{Put object}
\usage{
s3api_put_object(filepath, bucket, key, ...)
}
\arguments{
\item{...}{Additional AWS CLI s3api args, excluding `bucket`, `key`, and `body`.}
}
