% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list2args.R
\name{list2args}
\alias{list2args}
\title{Convert a list to AWS CLI args}
\usage{
list2args(...)
}
\arguments{
\item{...}{A list of arguments that will be mapped to command-line args for a system call to the AWS CLI utility.
Unnamed list entries are inserted into the system call as "<val>", i.e. with no argument-specifying hyphens.
Named list entries are inserted into the system call with the "--<name> <val>" pattern.
Any named list entries a \code{NULL} value will be inserted into the system call as "--<name>" only, i.e. as a named flag.
Arguments are constructed from the list in the same order as provided in the list.}
}
\value{
A character vector of arguments suitable for a \code{sys::exec_internal} call.
}
\examples{
list2args("foo", bar = "baz", flag = NULL)
}
